HIVE-27956: Query based compactor implementation separation
