+++ ql/src/java/org/apache/hadoop/hive/ql/txn/compactor/MajorQueryCompactor.java
@@ -72,10 +72,9 @@ final class MajorQueryCompactor extends QueryCompactor {
    * See {@link org.apache.hadoop.hive.conf.HiveConf.ConfVars#SPLIT_GROUPING_MODE} for the config description.
    */
   private List<String> getCreateQueries(String fullName, Table t, String tmpTableLocation) {
-    return Lists.newArrayList(new CompactionQueryBuilder(
+    return Lists.newArrayList(new CompactionQueryBuilderForMajor(
         CompactionType.MAJOR,
         CompactionQueryBuilder.Operation.CREATE,
-        false,
         fullName)
         .setSourceTab(t)
         .setLocation(tmpTableLocation)
@@ -84,10 +83,9 @@ final class MajorQueryCompactor extends QueryCompactor {
 
   private List<String> getCompactionQueries(Table t, Partition p, String tmpName) {
     return Lists.newArrayList(
-        new CompactionQueryBuilder(
+        new CompactionQueryBuilderForMajor(
             CompactionType.MAJOR,
             CompactionQueryBuilder.Operation.INSERT,
-            false,
             tmpName)
             .setSourceTab(t)
             .setSourcePartition(p)
@@ -96,10 +94,9 @@ final class MajorQueryCompactor extends QueryCompactor {
 
   private List<String> getDropQueries(String tmpTableName) {
     return Lists.newArrayList(
-        new CompactionQueryBuilder(
+        new CompactionQueryBuilderForMajor(
             CompactionType.MAJOR,
             CompactionQueryBuilder.Operation.DROP,
-            false,
             tmpTableName).build());
   }
 }
