+++ ql/src/java/org/apache/hadoop/hive/ql/txn/compactor/RebalanceQueryCompactor.java
@@ -73,10 +73,9 @@ final class RebalanceQueryCompactor extends QueryCompactor {
   }
 
   private List<String> getCreateQueries(String fullName, Table t, String tmpTableLocation) {
-    return Lists.newArrayList(new CompactionQueryBuilder(
+    return Lists.newArrayList(new CompactionQueryBuilderForRebalance(
         CompactionType.REBALANCE,
         CompactionQueryBuilder.Operation.CREATE,
-        false,
         fullName)
         .setSourceTab(t)
         .setLocation(tmpTableLocation)
@@ -85,24 +84,22 @@ final class RebalanceQueryCompactor extends QueryCompactor {
 
   private List<String> getCompactionQueries(Table t, Partition p, String tmpName, int numberOfBuckets, String orderByClause) {
     return Lists.newArrayList(
-        new CompactionQueryBuilder(
+        new CompactionQueryBuilderForRebalance(
             CompactionType.REBALANCE,
             CompactionQueryBuilder.Operation.INSERT,
-            false,
             tmpName)
-            .setSourceTab(t)
-            .setSourcePartition(p)
             .setNumberOfBuckets(numberOfBuckets)
             .setOrderByClause(orderByClause)
+            .setSourceTab(t)
+            .setSourcePartition(p)
             .build());
   }
 
   private List<String> getDropQueries(String tmpTableName) {
     return Lists.newArrayList(
-        new CompactionQueryBuilder(
+        new CompactionQueryBuilderForRebalance(
             CompactionType.REBALANCE,
             CompactionQueryBuilder.Operation.DROP,
-            false,
             tmpTableName).build());
   }
 }
